Add delete function and check button to update the dom
